fix(Keyboard): updates keyboard styles
