+++ src/packages/solid/components/Keyboard/Keyboard.styles.ts
@@ -16,29 +16,9 @@
  */
 
 import theme from 'theme';
-import type { Tone } from '../../types/types.js';
-import type { ComponentStyleConfig, NodeStyleSet, TextStyleSet } from '../../types/types.js';
 import { makeComponentStyles } from '../../utils/index.js';
-import type { NodeStyles } from '@lightningtv/solid';
-import type { KeySizes, KeyStyles, KeyConfig } from '../Key/Key.types.js';
-
-export interface KeyboardStyles {
-  tone: Tone;
-  Container: NodeStyleSet<{ keySpacing?: number; screenW?: number; marginX?: number }>;
-  Key: NodeStyleSet<{ baseWidth: number; sizes: KeySizes; contentColor: NodeStyles['color'] }>;
-  Text: TextStyleSet;
-}
-
-export type KeyboardStyleProperties = {
-  keySpacing?: number;
-  screenW?: number;
-  marginX?: number;
-  textColor?: NodeStyles['color'];
-};
-
-export type KeyboardConfig = ComponentStyleConfig<KeyboardStyleProperties>;
-
-// export type KeyConfig = ComponentStyleConfig<KeyStyleProperties>;
+import type { KeyStyles, KeyConfig } from '../Key/Key.types.js';
+import type { KeyboardStyles, KeyboardConfig } from './Keyboard.types.js';
 
 /* @ts-expect-error next-line themes are supplied by client applications so this setup is necessary */
 const { Keyboard: { defaultTone, ...themeStyles } = {} } = theme?.componentConfig;
@@ -47,12 +27,12 @@ const { Key: { ...keyThemeStyles } = {} } = theme?.componentConfig; // TODO defa
 
 const container: KeyboardConfig = {
   themeKeys: {
-    keySpacing: 'keySpacing',
+    gap: 'keySpacing',
     screenW: 'screenW',
     marginX: 'marginX'
   },
   base: {
-    keySpacing: theme.spacer.md,
+    gap: theme.spacer.md,
     screenW: theme.layout.screenW,
     marginX: theme.layout.marginX,
     height: 100
@@ -63,7 +43,7 @@ const container: KeyboardConfig = {
 
 const key: KeyConfig = {
   themeKeys: {
-    keySpacing: 'keySpacing',
+    gap: 'keySpacing',
     textAlign: 'textAlign',
     borderRadius: 'borderRadius',
     color: 'backgroundColor',
@@ -73,7 +53,7 @@ const key: KeyConfig = {
     contentColor: 'contentColor' // what is this used for
   },
   base: {
-    keySpacing: theme.spacer.md,
+    gap: theme.spacer.md,
     height: theme.spacer.md * 9,
     sizes: {
       sm: 1,
