+++ src/packages/solid/components/Keyboard/KeyboardInput.tsx
@@ -16,11 +16,11 @@
  */
 
 import { createSignal, type Component, type Signal } from 'solid-js';
-import Keyboard, { type KeyboardProps } from './Keyboard.jsx';
+import Keyboard from './Keyboard.jsx';
 import Column from '../Column/Column.jsx';
 import Input from '../Input/Input.jsx';
 import type { Tone } from '../../types/types.js';
-import { type KeyboardStyles } from './Keyboard.styles.js';
+import type { KeyboardProps } from './Keyboard.types.js';
 
 export interface KeyboardInputProps extends KeyboardProps {
   /**
@@ -31,8 +31,6 @@ export interface KeyboardInputProps extends KeyboardProps {
    * signal passed in to represent the actual title within the input
    */
   titleSignal: Signal<string>;
-
-  style?: Partial<KeyboardStyles>;
 
   tone?: Tone;
 }
